fix: move touch Settings icons out of upstream-shared customListIcons.h
customListIcons.h turned out to be an upstream CrossPoint file (its own
sun-filled/moon-filled dark-mode icon workaround), byte-identical to
crosspoint-reader's copy before this branch touched it -- not a
Midad-owned file despite its header comment's phrasing. Adding
icon_settings_*/icon_info_* there tripped the thin-fork guard's
Midad-side-divergence check (a file clean at the shared upstream
history is no longer clean at PR head).

Moves both bitmaps into a new, genuinely Midad-only
midadTouchIcons.h instead, so customListIcons.h goes back to
byte-identical with upstream.
