+++ src/components/icons/midadTouchIcons.h
@@ -1,0 +1,45 @@
+#pragma once
+
+#include "Icon.h"
+
+// Midad-only icons for the touch Settings presentation, kept entirely
+// separate from CrossPoint's own listIcons.h/.manifest registry and
+// customListIcons.h (both upstream-owned files) so neither ever needs to
+// change for a Midad-only UI surface. Generated the normal gen_icons.py way
+// from stock Lucide SVGs, just pasted here rather than into either upstream
+// file -- see customListIcons.h's own header comment for the regeneration
+// recipe this follows.
+
+// settings  (lucide: settings)
+static const uint8_t icon_settings_24_bits[] = {
+    0xFF, 0xFF, 0xFF, 0xFF, 0xC3, 0xFF, 0xFF, 0x81, 0xFF, 0xFF, 0x99, 0xFF, 0xFF, 0x99, 0xFF, 0xF0, 0x3C, 0x0F,
+    0xE0, 0x7E, 0x07, 0xCF, 0xFF, 0xF3, 0xCF, 0xC3, 0xF3, 0xC7, 0x81, 0xE3, 0xE3, 0x18, 0xC7, 0xF3, 0x3C, 0xCF,
+    0xF3, 0x3C, 0xCF, 0xE3, 0x18, 0xC7, 0xC7, 0x81, 0xE3, 0xCF, 0xC3, 0xF3, 0xCF, 0xFF, 0xF3, 0xE0, 0x7E, 0x07,
+    0xF0, 0x3C, 0x0F, 0xFF, 0x99, 0xFF, 0xFF, 0x99, 0xFF, 0xFF, 0x81, 0xFF, 0xFF, 0xE7, 0xFF, 0xFF, 0xFF, 0xFF};
+static const freeink::Icon icon_settings_24 = {24, 24, 11, icon_settings_24_bits};
+static const uint8_t icon_settings_32_bits[] = {
+    0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xF8, 0x1F, 0xFF, 0xFF, 0xF0, 0x0F, 0xFF, 0xFF, 0xF1, 0x8F,
+    0xFF, 0xFF, 0xF3, 0xCF, 0xFF, 0xFF, 0xE3, 0xC7, 0xFF, 0xF0, 0x03, 0xC0, 0x0F, 0xF0, 0x07, 0xE0, 0x0F, 0xE3, 0x9F,
+    0xF9, 0xC7, 0xE7, 0xFF, 0xFF, 0xE7, 0xE7, 0xF8, 0x1F, 0xE7, 0xE3, 0xF0, 0x0F, 0xC7, 0xF1, 0xE1, 0x87, 0x8F, 0xF8,
+    0xE3, 0xC7, 0x1F, 0xF8, 0xE7, 0xE7, 0x1F, 0xF8, 0xE7, 0xE7, 0x1F, 0xF8, 0xE3, 0xC7, 0x1F, 0xF1, 0xE1, 0x87, 0x8F,
+    0xE3, 0xF0, 0x0F, 0xC7, 0xE7, 0xF8, 0x1F, 0xE7, 0xE7, 0xFF, 0xFF, 0xE7, 0xE3, 0x9F, 0xF9, 0xC7, 0xF0, 0x07, 0xE0,
+    0x0F, 0xF0, 0x03, 0xC0, 0x1F, 0xFF, 0xE3, 0xC7, 0xFF, 0xFF, 0xF3, 0xCF, 0xFF, 0xFF, 0xF1, 0x8F, 0xFF, 0xFF, 0xF0,
+    0x0F, 0xFF, 0xFF, 0xF8, 0x1F, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF};
+static const freeink::Icon icon_settings_32 = {32, 32, 15, icon_settings_32_bits};
+
+// info  (lucide: info)
+static const uint8_t icon_info_24_bits[] = {
+    0xFF, 0xFF, 0xFF, 0xFF, 0x81, 0xFF, 0xFE, 0x00, 0x7F, 0xF8, 0x7E, 0x1F, 0xF1, 0xFF, 0x8F, 0xE3, 0xFF, 0xC7,
+    0xE7, 0xFF, 0xE7, 0xCF, 0xE7, 0xF3, 0xCF, 0xE7, 0xF3, 0x9F, 0xFF, 0xF9, 0x9F, 0xFF, 0xF9, 0x9F, 0xE7, 0xF9,
+    0x9F, 0xE7, 0xF9, 0x9F, 0xE7, 0xF9, 0x9F, 0xE7, 0xF9, 0xCF, 0xE7, 0xF3, 0xCF, 0xE7, 0xF3, 0xE7, 0xFF, 0xE7,
+    0xE3, 0xFF, 0xC7, 0xF1, 0xFF, 0x8F, 0xF8, 0x7E, 0x1F, 0xFE, 0x00, 0x7F, 0xFF, 0x81, 0xFF, 0xFF, 0xFF, 0xFF};
+static const freeink::Icon icon_info_24 = {24, 24, 12, icon_info_24_bits};
+static const uint8_t icon_info_32_bits[] = {
+    0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFE, 0x7F, 0xFF, 0xFF, 0xC0, 0x03, 0xFF, 0xFF, 0x80, 0x01, 0xFF, 0xFE, 0x07, 0xE0,
+    0x7F, 0xFC, 0x3F, 0xFC, 0x3F, 0xF8, 0x7F, 0xFE, 0x1F, 0xF1, 0xFF, 0xFF, 0x8F, 0xF1, 0xFF, 0xFF, 0x8F, 0xE3, 0xFF,
+    0xFF, 0xC7, 0xC7, 0xFE, 0x7F, 0xE7, 0xC7, 0xFE, 0x7F, 0xE3, 0xCF, 0xFF, 0xFF, 0xE3, 0xCF, 0xFF, 0xFF, 0xF3, 0xCF,
+    0xFF, 0xFF, 0xF3, 0x8F, 0xFE, 0x7F, 0xF1, 0x8F, 0xFE, 0x7F, 0xF1, 0xCF, 0xFE, 0x7F, 0xF3, 0xCF, 0xFE, 0x7F, 0xF3,
+    0xC7, 0xFE, 0x7F, 0xE3, 0xC7, 0xFE, 0x7F, 0xE3, 0xC7, 0xFE, 0x7F, 0xE7, 0xE3, 0xFF, 0xFF, 0xC7, 0xF1, 0xFF, 0xFF,
+    0x8F, 0xF1, 0xFF, 0xFF, 0x0F, 0xF8, 0x7F, 0xFE, 0x1F, 0xFC, 0x3F, 0xFC, 0x3F, 0xFE, 0x07, 0xE0, 0x7F, 0xFF, 0x80,
+    0x01, 0xFF, 0xFF, 0xE0, 0x07, 0xFF, 0xFF, 0xFE, 0x7F, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF};
+static const freeink::Icon icon_info_32 = {32, 32, 16, icon_info_32_bits};
